pyproject.toml
[project]
name = "a8p5-hands-on-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.9.0",
    "bitsandbytes>=0.46.1",
    "bs4>=0.0.2",
    "chroma-haystack>=3.3.0",
    "chromadb>=1.0.15",
    "einops>=0.8.1",
    "fastapi>=0.116.1",
    "haystack-ai>=2.16.1",
    "huggingface-hub[hf-xet]>=0.34.2",
    "hydra-core>=1.3.2",
    "ipykernel>=6.30.0",
    "ipywidgets>=8.1.7",
    "nltk>=3.9.1",
    "numpy==2.2",
    "omegaconf>=2.3.0",
    "openai-whisper>=20250625",
    "paddlepaddle>=2.5.0",
    "paddleocr>=3.1.0",
    "pillow>=11.3.0",
    "playwright>=1.54.0",
    "pydantic>=2.11.7",
    "pypdf>=5.9.0",
    "pypdf2>=3.0.1",
    "python-json-logger>=3.3.0",
    "python-pptx>=1.0.2",
    "requests>=2.32.4",
    "rouge-score>=0.1.2",
    "sentence-transformers>=5.0.0",
    "streamlit>=1.47.1",
    "timm>=1.0.19",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.54.0",
    "uuid>=1.30",
    "uvicorn>=0.35.0",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128" },
]
torchvision = [
  { index = "pytorch-cu128" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
